Finish 1, 2 of Stoch Experiments
Just need to add VolMAE. Results are so far how you would expect. There
is a complex dependence on TAU. Most of the dependence of the method
comes from the informativness of y. If y is very informative then we
should expect our method to perform better. Whereas when it is not then
this additional information has little extra value as opposed to the
boostrap.